Public fields in MavenCoordinate was replaced by setters and getters
Direct access to class fields has been replaced with access via getters and setters.
This solution provides limited control over the state of the class object and conforms
to the principles of encapsulation.

Added the testContext field.
The decision was made to separate the dependencies into "main" and "test".
This is necessary to properly add them to the configuration.
If a dependency in the project is declared as "testImplementation",
then after resolution it will be added to the same configuration as its transitive dependencies.