Remove class BomDependencyManager
The class was removed due to a revision of the logic for resolving BOM files.
Its logic has been partially transferred to the BomProcessor.
Now the dependencies specified in the BOM files are not added to the configuration after processing.
BOM files are used only for dependency versioning.
This helps to avoid unnecessary artifacts entering the configuration.