Create class PomHierarchyLoader to build a hierarchy of pom files
The class is responsible for building a chain of pom files (oldest parent -> youngest child).
This is necessary for the correct collection of data when parsing pom files, in particular the collection of "properties".
In turn, they are needed to correctly disclose placeholders and obtain a complete list of dependencies from the current POM.
The length of the chain is limited (10 elements) to prevent infinite recursion causing a "java heap".

Attention: all POM files must be in the same directory (parents and children).